주석해놓은거 해제함

+++ src/main/java/com/nhnacademy/_vidiabookstoreservice/user/service/impl/AddressServiceImpl.java
@@ -122,10 +122,10 @@ public class AddressServiceImpl implements AddressService {
         }
 
         // 삭제하려고 하는 주소가 기본주소인지 체크
-//        UserAddress defaultAddress = user.getUserAddress();
-//        if (defaultAddress != null && defaultAddress.getUserAddressId().equals(userAddress.getUserAddressId())) {
-//            throw new DefaultAddressCannotBeDeletedException("기본 주소는 삭제할 수 없습니다. 기본 주소를 변경 후 시도해주세요. 현재 기본 주소 별칭 : " + defaultAddress.getAlias());
-//        }
+        UserAddress defaultAddress = user.getUserAddress();
+        if (defaultAddress != null && defaultAddress.getUserAddressId().equals(userAddress.getUserAddressId())) {
+            throw new DefaultAddressCannotBeDeletedException("기본 주소는 삭제할 수 없습니다. 기본 주소를 변경 후 시도해주세요. 현재 기본 주소 별칭 : " + defaultAddress.getAlias());
+        }
 
         addressRepository.deleteByUser_UserIdAndUserAddressId(userId, addressId);
     }
